+++ Src/Pscx/PscxNouns.cs
@@ -13,19 +13,15 @@ namespace Pscx
         public const string Unit = "Unit";
         public const string Base64 = "Base64";
         public const string File = "File";
-        public const string FileTail = "FileTail";
         public const string FileVersionInfo = "FileVersionInfo";
         public const string Hash = "PscxHash";
-        public const string LoremIpsum = "LoremIpsum";
         public const string Object = "Object";
         public const string Script = "Script";
         public const string TypeName = "TypeName";
         public const string Xml = "Xml";
         public const string Yaml = "Yaml";
-        public const string MacOs9LineEnding = "MacOs9LineEnding";
         public const string UnixLineEnding = "UnixLineEnding";
         public const string WindowsLineEnding = "WindowsLineEnding";
-        public const string String = "PscxString";
         public const string ForegroundWindow = "ForegroundWindow";
 
         // FileSystem
@@ -34,7 +30,6 @@ namespace Pscx
 
         // Formatting
         public const string Byte = "Byte";
-        public const string Hex = "Hex";
 
         // EnvironmentBlock
         public const string EnvironmentBlock = "EnvironmentBlock";